+++ lib/src/backends/recursion_limiter.dart
@@ -4,18 +4,19 @@ import 'package:activatory/src/activation_context.dart';
 import 'package:activatory/src/backends/generator_backend.dart';
 import 'package:activatory/src/backends/generator_backend_wrapper.dart';
 
-class RecurrencyLimiter<T> implements GeneratorBackend<T>, GeneratorBackendWrapper<T>{
+class RecursionLimiter<T> implements GeneratorBackend<T>, GeneratorBackendWrapper<T>{
 
   final Type _type;
   final GeneratorBackend<T> _wrapped;
+  final ClassMirror _listMirror = reflectClass(List);
 
-  RecurrencyLimiter(this._type, this._wrapped);
+  RecursionLimiter(this._type, this._wrapped);
 
   @override
   T get(ActivationContext context) {
     if(context.isVisitLimitReached(_type)){
       var classMirror = reflectClass(_type);
-      if(classMirror.isSubclassOf(reflectClass(List))){
+      if(classMirror.isSubclassOf(_listMirror)){
         return new List() as T;
       }
       return null;