+++ lib/src/settings.dart
@@ -1,7 +1,0 @@
-class Settings{
-  final int maxStackSizePerType;
-
-  const Settings(this.maxStackSizePerType);
-
-  const Settings.defaults():this(3);
-}